% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{GMD}
\alias{GMD}
\title{Download and Filter Global Macro Data}
\usage{
GMD(year = NULL, quarter = NULL, country = NULL)
}
\arguments{
\item{year}{A numeric value representing the desired year (e.g., 2025). If NULL, the latest available dataset is used.}

\item{quarter}{A numeric value representing the quarter (1, 3, 6, or 9). If NULL, the latest available dataset is used.}

\item{country}{A character string specifying the ISO3 country code (e.g., "CHN"). If NULL, returns all countries.}
}
\value{
A data frame containing the requested data.
}
\description{
This function downloads a dataset from the Global Macro Database based on the provided year, quarter, and country.
}
