% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMD.R
\name{GMD}
\alias{GMD}
\title{Download Global Macro Data}
\usage{
GMD(year = NULL, month = NULL, country = NULL)
}
\arguments{
\item{year}{A numeric value representing the desired year (e.g., 2025). If NULL, the latest default dataset is used.}

\item{month}{A numeric value representing the month (1, 4, 7, or 10). If NULL, the latest default dataset is used.}

\item{country}{A string specifying the ISO3 country code (e.g., "CHN" for China). If NULL, returns all countries.}
}
\value{
A data frame containing the requested data.
}
\description{
This function downloads a dataset from the Global Macro Database.
If no year and month are provided, it fetches the default dataset from "GMD.csv".
}
