% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{gmd}
\alias{gmd}
\title{Get macroeconomic data from Global Macro Data}
\usage{
gmd(version = NULL, country = NULL, variables = NULL, show_preview = TRUE)
}
\arguments{
\item{version}{A string representing the dataset version (e.g., "2025_01").}

\item{country}{A string or vector of ISO3 country codes (e.g., "USA" or c("CHN", "USA")).}

\item{variables}{A vector of variable names to include.}

\item{show_preview}{A boolean indicating whether to show a preview.}
}
\value{
A dataframe containing the requested macroeconomic data.
}
\description{
Get macroeconomic data from Global Macro Data
}
